Global: overall optimization

+++ src/fastfetch.h
@@ -27,11 +27,11 @@
 
 #ifdef __has_builtin
     #if __has_builtin(__builtin_types_compatible_p)
-        #define ARRAY_SIZE(x) ({ static_assert(!__builtin_types_compatible_p(__typeof__(x), __typeof__(&*(x))), "Must not be a pointer"); sizeof(x) / sizeof(*(x)); })
+        #define ARRAY_SIZE(x) ({ static_assert(!__builtin_types_compatible_p(__typeof__(x), __typeof__(&*(x))), "Must not be a pointer"); (uint32_t) (sizeof(x) / sizeof(*(x))); })
     #endif
 #endif
 #ifndef ARRAY_SIZE
-    #define ARRAY_SIZE(x) (sizeof(x) / sizeof(*(x)))
+    #define ARRAY_SIZE(x) ((uint32_t) (sizeof(x) / sizeof(*(x))))
 #endif
